% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry_exit.r
\name{EntryExit}
\alias{EntryExit}
\title{Firm entry and exit status}
\usage{
EntryExit(geo_dta, harmonized, base_year, years)
}
\arguments{
\item{geo_dta}{Input cleaned geographic data, stacked over years.}

\item{harmonized}{Input True if the input geo_dta has firm ID harmonized.}

\item{base_year}{A vector of base years to which firms from future survey years are compared. Default is 2001}

\item{years}{A vector of survey years that need firm status relative to the base_year, should include the base_year}
}
\value{
Either a stored data in store_dir, or a cleaned data frame.  A data frame with  rows and  variables
}
\description{
Input geo_data and
get firm entry, exit, incumbent status
}
\details{
The list of data dn has to be ordered in an ascending order.
}
