% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_firmID.R
\name{harmonize_firmID}
\alias{harmonize_firmID}
\title{Harmonize firm identifiers to create unique identifiers for panel firms from 2001-2018}
\usage{
harmonize_firmID(geo_dta)
}
\arguments{
\item{geo_dta}{geographic firm data. See getLocation function.}
}
\value{
A cleaned data frame with year, geographic location (not harmonized), sector (not harmonized),
unique tax id, and unique firm id
}
\description{
Input geographic data, stacked across years.
}
\details{
The input data should include years 2015 to 2017 where most changes happen for firm IDs
}
