% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-inputData.R
\name{InputData}
\alias{InputData}
\title{Copy Raw Firm Data}
\usage{
InputData(dta_dir)
}
\arguments{
\item{dta_dir}{A path to local location of raw data at year level}
}
\value{
A copied version of the folder in inst/extdata
}
\description{
Get links to Firm data folders and copy to package
}
\examples{
InputData("/Volumes/GoogleDrive/My Drive/econ_datasets/Vietnam_VES/Data/Stata_2007_2009/Stata_2007")
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
