% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skill_07_dta}
\alias{skill_07_dta}
\title{Firm 2007 Skill Intensity}
\format{
A data frame with 155770 rows and 9 variables:
\describe{
\item{\code{svyear}}{double Survey year}
\item{\code{tinh}}{integer Province of location}
\item{\code{macs}}{integer Plant ID}
\item{\code{madn}}{integer Firm ID}
\item{\code{ma_thue}}{integer Tax ID}
\item{\code{total_L}}{double Total number of employees at the end of the year}
\item{\code{SI_1}}{double (PhD + Masters + Uni + College + skilled workers)/ total_L) where skilled workers are professional and long term vocationally trained}
\item{\code{SI_2}}{double The same as SI_1 except dropping college (Cao Dang)}
\item{\code{SI_3}}{double The same as SI_2 except dropping high_skill_tec}
}
}
\source{
\url{https://www.gso.gov.vn/en/enterprises/}
}
\usage{
skill_07_dta
}
\description{
Firm skill intensity measures in 2007, measured by labor composition at the end of 2007
}
\keyword{datasets}
