% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wage_dta}
\alias{wage_dta}
\title{Wage data}
\format{
A data frame with 212322 rows and 8 variables:
\describe{
\item{\code{tinh}}{character COLUMN_DESCRIPTION}
\item{\code{macs}}{double COLUMN_DESCRIPTION}
\item{\code{madn}}{double COLUMN_DESCRIPTION}
\item{\code{ma_thue}}{character COLUMN_DESCRIPTION}
\item{\code{total_L}}{double Total number of employees at the end of the year}
\item{\code{wage_bill}}{double Compensation for employees including wages, salaries, bonus, social security, and other compensation out of production costs}
\item{\code{ins_pen}}{double Employer's contributions to insurance and pension, health, trade-union}
\item{\code{svyear}}{double survey year}
}
}
\source{
\url{https://www.gso.gov.vn/en/enterprises/}
}
\usage{
wage_dta
}
\description{
Firm compensation for employees
}
\keyword{datasets}
