% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-clean_skill.R
\name{clean_skill}
\alias{clean_skill}
\title{Clean and create skill intensity variables}
\usage{
clean_skill(dta)
}
\arguments{
\item{dta}{raw skill composition data with tinh, madn, macs, ma_thue, svyear variables}
}
\value{
Cleaned data frame with skill intensity measures
}
\description{
Get firm characteristics including skill composition
and total labot and output firm skill intensity measures
}
\details{
DETAILS
}
