% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry_exit.r
\name{EntryExit}
\alias{EntryExit}
\title{Firm entry and exit status}
\usage{
EntryExit(
  dta_list,
  store_dir,
  base_year = c(2001),
  years = c(2001, 2004, 2007)
)
}
\arguments{
\item{dta_list}{Input a list of enterprise dn files from GSO}

\item{store_dir}{If provided a store_dir, then the output wage data frame will be stored there.
Otherwise, output the cleaned data frame.}

\item{base_year}{A vector of base years to which firms from future survey years are compared. Default is 2001}

\item{years}{A vector of survey years that need firm status relative to the base_year, should include the base_year}
}
\value{
Either a stored data in store_dir, or a cleaned data frame.  A data frame with  rows and  variables
}
\description{
Input a list of enterprise dn files and
get firm entry, exit, incumbent status
}
\details{
The list of data dn has to be ordered correctly, first 2001, 2004 then 2007.
}
