% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm_location.r
\name{getLocation}
\alias{getLocation}
\title{Firm geographic location 2001 and 2007, Not harmonized}
\usage{
getLocation(dta_list, store_dir)
}
\arguments{
\item{dta_list}{raw data list dn from GSO}

\item{store_dir}{If provided a store_dir, then the output wage data frame will be stored there.
Otherwise, output the cleaned data frame.}
}
\value{
Either a stored data in store_dir, or a cleaned data frame.  A data frame with  rows and  variables
}
\description{
Input a list of enterprise dn files and
get location data for firms between 2001 and 2007
}
\details{
The list of data dn has to be ordered correctly, first 2001 then 2007.
}
