% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWage.R
\name{getWage}
\alias{getWage}
\title{Get wage bill data for firms 2001 and 2007}
\usage{
getWage(dta_list, store_dir)
}
\arguments{
\item{dta_list}{raw data list dn from GSO}

\item{store_dir}{If provided a store_dir, then the output wage data frame will be stored there.
Otherwise, output the cleaned data frame.}
}
\value{
Either a stored data in store_dir, or a cleaned data frame.  A data frame with 212322 rows and 8 variables
}
\description{
Input a list of enterprise dn files and
get wage bill data for firms between 2001 and 2007
}
\details{
The list of data dn has to be ordered correctly, first 2001 then 2007.
Firm compensation for employees, already removed missing values for total_L, wage_bill, and ave_wage, and firms with zero employees.
A data frame with 212322 rows and 8 variables
}
