% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownership_type.r
\name{ownership}
\alias{ownership}
\title{Firm harmonized-ownership type 2001-2019}
\usage{
ownership(geo_dta)
}
\arguments{
\item{geo_dta}{geographic data 2001-2019}
}
\value{
Either a stored data in store_dir, or a cleaned data frame.  A data frame with  rows and  variables
}
\description{
Input a list of enterprise dn files and
get location data for firms 2001, 2004 and 2007
}
\details{
The list of data dn has to be ordered correctly to match with the years vector of survey years.
The harmonization follows Mccaig et al (2020) Appendix Table 4: Ownership types by year.
}
